refactor: consolidate openai utilities
