+++ frontend/packages/shared/src/api/index.ts
@@ -1,3 +1,2 @@
 export * from './client';
 export * from './auth';
-export * from './openai';
